Allow OIDC with Azure B2C
There are a couple of places where the client_secret should not be passed to Azure B2C as part
of the authentication requests. B2C is picky about this, and will reject the token grant if passed.
The data bindings also needed to be improved and allowed to suit custom values, as the B2C data bindings
are different to the default bindings in AAD.